First commit.
SetCPUGov is a simple SHell script that has been cobbled together in about a minute or less to make it easier for me to change CPU governors when needed. Yeah, I know I could make cpufreq do it with a system service, but sometimes you need action NOW! ... and that is what this is for. :-)